fix(account-lib): fix decode signed algo transaction
Algosdk has a issue with its type system right now where the
decoded sign transaction is of type EncodedTransaction when it
is actually of type Transaction. This fix is a temporary work
around until the issue is resolved.

See the github link in the comments for more context.

ticket: BG-31558