fix(core): defer application of authorization headers
Prior to v12, the `Authorization` header was applied just before sending a
request, but in v12 it was applied when the request was instantiated. It's
possible for tokens to be swapped out in time between the request getting
instantiated and getting sent, so we need to continue to lazily apply the
`Authorization` header.

Ticket: BG-34772