fix(core): fix regression in `addAccessToken` when using v1 auth
Prior to v12, we always called `handleResponseResult()` on the value returned
from the result of the call to platform, whether it was a v1 or v2 request.

After a bad refactor in v12, `handleResponseResult()` was no longer called when
using v1 auth. This broke callers who relied on this behavior for v1.

Instead, call `handleResponseResult()` on both return paths.

Additionally, add in some debug lines which were helpful in diagnosing this
issue.

Ticket: BG-34772