fix(core): fix issue while signing eos transaction using OVC
During uploading an unsigned EOS tx file in OVC, an issue was coming "coin undefined is not defined"
because field "coin" was expected in the recovery file but it was not being added while forming the
recovery transaction. I have added the coin field. While signing the unsigned EOS tx using EOS, an
issue was coming because signTransaction function was expecting a field txHex, but this field was
not there in the unsigned transaction. So I have added txHex field as well in unsigned recovery tx
file

"fix 30012"