chore(utxo-lib): adjust coverage parameter
The previous commit removed a bunch of code that was "covered" but
never really used in production. As a result, the coverage _ratio_
has decreased a tiny bit.

Sadly adding the manual annotation for the exception paths alone did
not push the ratio above 89 so we have to shave off another point.

Issue: BG-35119