fix(account-lib): readd `es5` target and `esModuleInterop`
These turn out to be needed for the tests and import styles used in
account-lib. While these should be changed to be compatible with more
modern targets, that's beyond the scope of this PR.

Also disable `esModuleInterop` at the root level again, since we really
only need this for the `account-lib` module.

Ticket: BG-34797