chore(root): remove `commitizen`
We already check commit message formats during drone builds, so this
extra local check seems to just get in the way and confuse new
developers.

Let's remove it and rely instead on the commit message format checks in
drone.

If this makes for too slow an iteration cycle, we can instead run the
commitlint check in a pre-push script. We can revisit this in a later
PR.

Ticket: BG-35311