fix(core): handle script sigs without signature property in `explainTransaction`
Some bitcoin forks use single signature p2sh inputs for replay protection.

When these inputs are parsed by `parseSignatureScript` in utxo-lib, the
returned object won't have a `signatures` property, since it's not recognized
as a 2-of-3 p2sh input. This is unique to replay protection inputs.

This fixes the `explainTransaction` method to handle parsed sig scripts that
don't have a `signatures` property by falling back to an empty array if this
property is not defined.

Ticket: BG-36155