fix(root): resolve `axios` to `^0.21.2`
This temporarily resolves the `yarn audit` failures, but we should revisit to
actually update the `axios` dependency versions used by dependencies. If our
dependencies are well behaved, we should just be able to remove the resolution
and make sure the `yarn.lock` is updated to point to a patched version of
`axios`.

Ticket: CR-400