fix(utxo-lib): fix sighash for dash transactions
Previously, we were not including `extraPayload` in the hash preimage
for transaction signatures. This was because the `hashForSignature`
method calls the `__toBuffer()` method which was not overridden in the
DashTransaction subclass (only the public `toBuffer()` method was).

This change now overrides `__toBuffer` with a not so pretty hack. We
can fix that later.

Test fixtures were created using utxolib `v1.9.6`.

Issue: BG-37672