build(core): bump @bitgo/unspents dependency
This upgrades to the latest @bitgo/unspents dependency which introduces
support for p2tr key path and script path input types. It also has
subtle changes to the way fees/dimensions are calculated for previously
supported input types, which required a number of changes to existing
test cases and fixtures to match the latest logic.

Ticket: BG-35578