feat(core): add support for p2tr script path sign
This adds support for p2tr script path signing for btc via the
`prebuildAndSignTransaction` interactive flow with the BitGo API. A test
case is added for p2tr transaction signing starting with a prebuilt,
unsigned transaction using a single p2tr input and then signing it twice
with the user and backup keys via a tapscript embedded in the taptree.
The resulting signed txHex represents a valid bitcoin transaction that
has been broadcast and confirmed on the live testnet btc chain.

Ticket: BG-35578