fix(statics): ensure UnderlyingAssets values are unique
each value in UnderlyingAsset represents an asset and clients use
the value to determine which icon to show.

additionally, the same underlying asset may be used by multiple chains
(e.g. USDT exists as an erc20 and eos token).

previously, we had two enum entries with the same value 'box'.
one was meant for the erc20 token and the other for eos.
this led to us showing the wrong icon for the eos box token.
fix is to make sure BOX and EOS_BOX have different values.