fix: exclude ripple-address-codec 4.2
The `ripple-address-codec` package published 4.2.0 2 days ago, which
included a breaking change of requiring npm >= 7. This is not the
default version of npm installed with node 14 and is a significant
change in behavior from npm 6, so not all clients can be expected to
update easily.

Therefore we pin `ripple-address-codec` to `~4.1.3`, the last-published
4.1 version.