feat(utxo-lib): support import from `src/bitgo`
While one could import from `@bitgo/utxo-lib/src/bitgo` prior to this
change, one would actually import the Typescript files instead of the
Javascript files in `dist/`.

This had unintended consequences since for instance the networks
defined in the Typescript files and the Javascript files are not
reference-equal and therefore cause `unknown network` errors.

With this change the import is resolved to the correct file.

Issue: BG-38773