fix(core): add base `explainTransaction` method
The method, while optional, is part of the generic coin interface.

Unfortunately it is very difficult to have very strong types here so
we have to add a bunch of type parameters.

The `options` argument in fact is so heterogeneous between coin
implementations that it cannot be narrowed to anything meaningful
besides `any` (see implementation for EOS).

Issue: BG-40467