feat(core): add support for user-provided custom signing function
This change allows a user to pass an arbitrary function as the
`customSigningFunction` property to any function which signs
transactions. This custom signing function will take all the same
arguments which are passed to a coin implementation's `signTransaction`
function, except no private key will be provided - the custom signing
function is responsible for acquiring and applying the user's private
key. It is expected to resolve to a valid half-signed transaction that
will then be sent to BitGo for cosigning via the normal `/tx/send`
route.

This effectively allows the user to bypass the built-in signing flow and
use their own implementation for transaction signing, while still
leveraging BitGoJS to verify the transaction and send it to BitGo for
cosigning.

Ticket: BG-41943