chore(account-lib): troubleshoot dot test OOM
Given that the SDK tests are run in parallel, all tests share
the same memory resource. In dot tests, the metadata rpc
variable has very heavy usage of memory resources given its
size, but we seem to duplicate it too much in the unit tests
in dot.

In this PR, all unnecessary usage that involves the metadata
rpc is refactored through:
* Only use the required variables through deconstructing
objects instead of copying the whole
* Making the config variable ephemeral so that this huge
variable is eligible for garbage collection

Comparing to:
* In the previous tests, all duplicated metadata was persisted
at the top level in the test suites for easy use of its value.

Ticket: BG-43065