fix(core): fix bip32/ecpair, API vs Interface
* bump `bip32` package to v3
* use `bip32: BIP32API` type from `@bitgo/utxo-lib`
* use `bip32: BIP32Interface` type from `bip32` package

* use `ecpair` package to v2.0.1
* use `expair: ECPairAPI` type from `@bitgo/utxo-lib`
* use `ecpair: ECPairInterface` type from `ecpair` package

This has the unfortunate downside that we are depending on
`@bitgo/utxo-lib` for bip32 again, after we removed this dependency in
earlier commits.

We should fix this by instantiating `bip32: BIP32API` per package.
However this is not easy to do here since we are depending on a
slightly exotic secp256k1 library in `@bitgo/utxo-lib` and I don't
want to depend on that directly in `BitGoJS` yet.

Issue: BG-41154

Co-authorered-by: brandonblack@bitgo.com