test(utxo-lib): add test  for high-S ECDSA signatures
Add `verifySignature()` test where we take an existing signature and
tweak the signature `S` value to `n - S` (high-S).

The resulting signature is valid according to old consensus rules, but
invalid according BIP-0146 which requires a low-S value.

https://github.com/bitcoin/bips/blob/master/bip-0146.mediawiki

Note that the test will fail in this commit, showing that we currently
do not reject high-S signatures. The fix is included in the next
commit.

Issue: BG-44787