chore: use keyType instead of type for tss + bls keys
type was originally used to differentiate different types of backup keys
(i.e. using an external key provider or holding on to your own).

this was overloaded with the introduction of TSS and BLS.

this change refactors to use keyType to determine multisig type (tss,
bls, or independent/onchain).

Ticket: BG-43890