feat(bitgo): support chaincodes on BLS-DKG keychains creation
As part of the BLS-DKG HS implementation, chaincodes needs to be supported at the time of keychain
creation so then it is possible to derive the subkeys when signing tx. Each participant contributes
with its own chaincode and those are added up to the common keychain.

Also fixes BG-47812. Add padding to the secret shares hex so it always
is 64. It sometimes was 62.
BREAKING CHANGE: This breaks the current ETH2 Hot Wallet creation flow. Needs BG-46182 to be
implemented and deployed too.

BG-46184