fix(bitgo): update the hardcoded decimal place to be dynamic
CHEX token is having issue displaying bignumber decimals where we
output 4 digits of decimals while they output 8 decimals. This is
due to us hardcoding the digits while it is already dynamic on the
tokens side.

Adding a default `decimalPlaces` field in the base class and call
it in `baseUnitsToBigUnits`.

Ticket: BG-47340