fix(bitgo): Add missing examples and filters for list addresses API
List addresses API supports the following query parameters to filter the response.

includeBalances - Whether to include address balances

includeTotalAddressCount - For large wallets (>100k addresses), include total count of addresses (including addresses pending on-chain) matching the query.

returnBalancesForToken - Name of the token that the response should include balances for. (Eth and Celo only)

pendingDeployment - Filter the addresses based on their deployment status. Return the deployed addresses if this param is passed as false and return undeployed addresses if it is passed as true. (Eth only)

But currently the SDK method ‘addresses’ which calls the API is not supporting the mentioned filters. So, this fix adds the missing filters to the SDK method along with unit test changes.

Also examples for the addresses SDK method are missing. So they are also added along with pendingDeployment, includeBalance & returnBalancesForToken filter usage example. These 3 filters example is for those who want the use case from  BG-37593.

Ticket: BG-39897