chore: notify yarn version in readme
There isn't a good way to enforce package manager version to be
yarn v1. There are two possible ways to specify the version: one
is through the `engine` option, and another is the `packageManager`
option in `package.json`.

Now with the `engine` configuration, it only notify the user of the
SDK that it is compatile with `yarn` v1 installs, but for developers
of the SDK, it won't enfore the package manager version.

With `packageManager`, you can specify like this:
```
...
  "packageManager": "^yarn@1.22.19"
...
```
But if you would like to use yarn v2 on this repo with:
```
yarn set version berry
```
It will override this config and the install will not be blocked either,
let alone that the option only works after `v16.9.x` without corepack,
which is not a dependency of the SDK.

Ticket: BG-00000