fix(sdk-coin-eos): fix precision for EOS:CHEX tokens
Currently, when withdrawing CHEX tokens on EOS mainnet, the amount is shown with 4 decimals of
precision on the FE. However, the precision should be 8 decimals for CHEX specifically.
This fix includes the modification of baseUnitsToBigUnits in the Eos class, which now takes in an
optional boolean flag that indicates if the EOS coin is CHEX and rounds the value appropriately.

Ticket: BG-47340