feat(sdk-core): tss ecdsa key creation flow with 3rd party backup
This modifies the ecdsa TSS flow to support third party backup key
providers (for Nike, BitGo is the backup provider) where the backup key
is created via Bitgo (i.e WP->HSM). It involves two additional calls
to the WP, first to create the backup shares, and second time to
update the backup key in WP with the bitgoToBackup shares and verifying
the common keychain. Rest of the flow remains the same.

Ticket: BG-59055