feat(utxo-lib): remove dependency on `classify` in parseSignatureScript
Our current script parsing code consists of

- decompiling a script
- checking the classification
- checking the length
- picking certain parts and validating them
- potential recursion for nested scripts

Doing this manually with hand-written conditionals and destructuring
is hard to read and error prone. Further we do not want to depend on
the `classify` methods any longer.

This change introduces a simple recursive pattern matching
implementation that can deal with the handful of scripts we need to
work with. It only depends on `opcodes` and `script.decompile`.

As a side benefit, the new implementation reduced the execution time
by ~5s on my machine.

Issue: BG-57748