chore(sdk-core): seperate ecdsa/eddsa create keychain typed params
Currently we are implementing create keychain methods for
eddsa and ecdsa based on their type definitions in base utils.
However, it's getting hard to maintain as the params and their
use case is deviating between ecdsa and eddsa.

This creates a new type for the common fields and lets ecdsa/eddsa
add new typed params of their own.

Ticket: BG-63193