fix(root): patch CVE-2022-46175
`lerna` and `eslint-plugin-import` both imported `tsconfig-paths`
that uses a vulnerable version of `json5`. While `lerna` patched
the fix in the latest release already, the later has not pushed any
fixes at this point yet.

Considering the `lerna` version that was patched is one major version
ahead of our current lerna version, `json5` is added for resolution
instead of a version update.

Ticket: BG-00000