fix(utxo-lib): fix zcash psbt from buffer
when we were deserializing a zcash psbt from a buffer, we were not removing
the consensusBranchId from the global map, so if we were to reserialize a
psbt that was deserialized from that buffer, we would have errored out
when trying to reserialized because the key was already in the global
map. By deleting it from the global map when we deserialize, we fix this
from happening.

BG-54756