feat(sdk-coin-dot): propagate keepAlive flag in tx build from hex
We currently only recognize transferAll transactions to have
the keepAlive flag as set to true. This is not always the case.

This commit updates the logic to check the transaction method
data when rebuilding the tx rather than always setting keepAlive
to true by default.

Also adds a unit test to ensure different txid's are crafted
in the keepAlive true/false cases.

TICKET: BG-65619