feat(statics): add new coin features for custody
As part of this PR, we're introducing four new coin features, one for
each of the BitGo entities. These new coin features will be used to
indicate if a coin supports custody in the associated BitGo entity or
not. The old Custody feature has been tagged as depreciated and
it's usage has been removed from the package. Moreover, the default
Custody feature of different types of coin groups has also been removed

BG-66027