feat(statics): add custody coin features for avaxc
We recently introduced new custody coin features where every coin has a
feature for all the BitGo entities it belongs to. We're now enforcing
these features in Wallet Platform and during this change, it came about
that avaxp is currently missing all the custody flags. As part of this
PR, we're adding the default BitGo Trust custody flag for avaxp coin.