feat(key-card): support BitGo Trust holding TSS backup key
The recovery process for when BitGo Trust holds the backup key for
TSS wallets looks a bit different from the normal recovery flows.
The purpose of the key card for TSS wallets where BitGo Trust holds
the backup key is for the scenario where BitGo Inc goes out of
business. In this situation BitGo Trust will publish the encrypted
backup seed. Users will then be able to combine it with the two
key shares on their key card (from user and from BitGo to backup).

Ticket: BG-64432