feat(statics): add expiring tx CoinFeature
Adds a CoinFeature called expiring_txs to statics for coins like DOT

CLOSES TICKET: BG-67958

We currently have coin specific logic in the trust admin tool, which is a no no.

The logic involves DOT, which has expiring transaction times.

Instead of using coin specific logic, the best path forward is to add a coinfeature in statics.

This PR adds an expiring txs coinfeature, and adds that feature to DOT.