feat(utxo-ord): use stepwise solution algorithm for inscription tx
Replace incomplete constraint resolution algorithm with a more
hand-crafted solution.

This now correctly builds transactions for inscriptions that are in
the middle or close to the end of an input.

Also remove the `maxChangeOutput` parameter since practically we don't
care much about the maximum size of change outputs for inscription
passing transactions.

Issue: BG-68315