feat(utxo-lib): improve type signatures for `createTransactionFrom*`
The new signatures make sure that the passing `amountType` results in a narrower
return type. The result of
`createTransactionFromBuffer(tx, network, { amountType: 'bigint' })` is now
`UtxoTransaction<bigint>` instead of `UtxoTransaction<number | bigint>`.

Mark `createTransactionFromHex` as deprecated. It is only a wrapper around
`createTransactionFromBuffer` and we need to repeat the extra type declarations.
It also does not have the `version` parameter, making it inconsistent with
`createTransactionFromBuffer`. It is better to handle the decoding at the call
site.

Issue: BTC-0