chore(bitgo): add .prettierrc.yml
The toplevel `.prettierrc.yml` has `requirePragma: true` which means it
does not apply to most of the files in the bitgo module.

I have seen multiple PRs with incidental formatting changes, probably because
the IDE does not pick up the `.prettierrc.yml` in the bitgo module.

This PR adds a `.prettierrc.yml` to the bitgo module but does do the
reformatting yet. That will be in the next PR.

To verify the changes, check out the PR and run these instructions

```
git checkout BTC-0.prettier-bitgo
git reset --hard HEAD~
yarn prettier --write .
git diff origin/BTC-0.prettier-bitgo # should be empty
```

Issue: BTC-0