feat(sdk-core): add postWithCodec utility function
This is a helper method for the development of request codecs.

Since it is difficult to get the codec right the first time, this method
allows to send the request with the original body and set some header
flags to indicate that the codec failed to decode or encode the body.

Issue: BTC-368