refactor(abstract-eth): move methods to abstract-eth WIN-1012
BREAKING CHANGE: Type of nextContractSequenceId field in TransactionPrebuild
interface is changed from string to number in AbstractEthLikeCoin and AbstractEthLikeNewCoins classes.
getCustomChainName method is removed from Polygon class because a common
method getCustomChainCommon has been added to AbstractEthLikeNewCoins
class for all EthLike coins. replayProtectionOptions is not optional in buildTransaction method in AbstractEthLikeNewCoins
and needs to be passed to derive the Eth common object from the chainId.
signFinalPolygon method name from Polygon class is updated to signFinalEthLike so that
it can be used for other EthLike coins. getBaseFactor method in Eth
and Polygon class returns number instead of string just to align with
AbstractEthLikeCoin
Ticket: WIN-1012