fix(root): improve error handling for consolidateAccount
Using the wallet method sendAccountConsolidations if the wrong password was provided there was no
check for it, so now before building and trying to sign the tx, it will fail earlier. Also when the
buildConsolidation returned an empty array(meaning no receive address has balance to consolidate) we
return an error.

WP-1034

TICKET: WP-1034