fix(sdk-api): fix proxy-agent usage
We removed the superagent-proxy earlier to use proxy-agent.
The change missed some old usage.

Change the proxy-agent version to 5.0. This is the version
that was used under the hood. The new version requires setting
proxy url in the environment variable. It does not quite fit how
we use the proxy-agent currently.

Ticket: WP-1185