fix(sdk-coin-hbar): hbar transaction valid duration
Ticket: WIN-1224

The client is unable to sign the cold wallet transactions within 120s.
They have requested to extend the window to 180s which is valid as per HBAR network.
This will give them some time to sign transactions while we work on a permanent fix to address this gracefully.