feat(bitgo): add PATCH and OPTIONS to redirectRequest
Currently if someone tries to use BitGo Express to do
a PATCH or OPTIONS call, we will fail because the code
we have does not have a path for either of these HTTP
Methods.

This PR adds paths for both methods, which should allow
users to be able to call endpoints like EVS's PATCH
identity endpoint via BitGo Express.

Ticket: COPS-1288