feat: add example for v1 sweep transaction flow
This commit adds a working example for sweeping from a v1btc wallet
given a list of unspents. A valid full signed transaction has been
generated and stored in `sweep-tx` file for future reference. Naturally,
this has not been broadcast in the `testnet` env as expected by the v1
sweep flow.

BTC-1224