fix: remove no-redeclare rule from eslint config
This rule yields false positives when exporting a type and a const with the same
name (as common in io-ts codec and type pairs).

The rule is checked by the typescript compiler, so it is not necessary to have
it in the eslint config.

See https://typescript-eslint.io/rules/no-redeclare/

Issue: BTC-1472