feat(utxo-lib): add toPsbtBuffer
First checks if the input is already a buffer that starts with the magic PSBT
byte sequence. If not, it checks if the input is a base64- or hex-encoded string
that starts with the PSBT header.

This function is useful when reading a file or request that could be in any of
the above formats.

Issue: BTC-1351