chore(sdk-core): preprocess build params before submitting
In the SDK we handle OP_RETURN outputs by setting them
in the address field of the output with the format
`scriptPubkey:HEX_ENCODED_SCRIPT_PUBKEY`. When we submit
to wallet platform, we want to convert that output to
the format { script: string, amount: string | number }.

Before submitting the buildParams to the API, we want
to convert it. In every other instance for other coins
we just want to return the same parameters.

TICKET: BTC-1582