feat(sdk-coin-apt): send many: addition of recipients
TICKET: COIN-3327

This is part of PRs for send many support for Aptos.
In this PR, adding new array of recipients but currently its having a single recipient in it.
Along with the maintaining backward compatibility for recipient flow.
In upcoming PRs, would be incrementally changing recipients to have batch support and multiple recipients.
And after successful migration, removing the recipient field and flow.