feat(utxo-lib): extract address from paygo attestation proof
Made a function that extracts the address buffer from an
attestation format. We created a testutil function that generates
this attestation format so that we can test within utxo-lib.

This messages will later be signed by a key and we need to verify
that signature as well as make sure that the address encoded in
the attestation proof matches the address of the output that we
attach this onto.

TICKET: BTC-2150